Bare-bones implemntaton of metadata auto upload.
add a new config file parameter ``md_mode'' that can we either ``local'' or
``remote''.

local acts as everything did before. remote assumes all md is stores on the
server (with a local cache that is not yet implemented). extracts and lists will
download the md file, then do the extract. archive will upload the md file when
done.

Caveats with the current unfinished implementation:
- md files are still made in the current directory or fullpath provided instead
of in a cache dir.
- we don't use the cache dir at all.
- we don't remove the files when done (so you'll leave md files around).
- md is currently not fully integrated into the event loop, so it starts the
loop again to do the up/download (meaning another login session etc). This is
planned to change eventually.
- probably many other assumptions.

``get it in and work on it in tree'' marco@. ``ok by me'' drahn@.

No documentation as this isn't really ready for consumption yet.