In remote md mode store and search for filenames in the format
YYYYMMDD-HHMMSS-<strnvis(user-supplied-tag>

extract defaults to extracting the newest backup with that tag. Code to
detect if a full timestamp has been provided is here but non-functional
(returns 0 when checking always) this will be fixed soon.

Remote mode is still in heavy development and should not be used quite yet.