Initial code for remote-mode directory cache.
If in remote mode, then the md_cachedir config file directive is
mandatory. All metadata will be created in the cache before being
uploaded and will be downloaded to the cachedir.

Currently the cache size is unlimited, this will be configurable soon.
remote mode also does not currently support differential backups, this
will also be fixed soon.