Begin tigher statemachine integration of metadata.
- instead of switching on ct_metadata in ct_process_complete set up the correct
functions in set_wakeup_write for the operation we are doing (extract or
archive). Switch in these functions on the C_HDR_F_METADATA flag in the heaer on
whether to treat is as metadata operations or normal file operations.

- unify the different transaction id counters to be one global one and one
global expected packet id.

- ct_metadata is no longer a global. just a local variable in main and does not
need to be set around md operations.

- one caveat is that due to md_list and md_erase not actually calling
completions right now we leak those packet ids in the statemachine, fix this up
in ct_shutdown until I manage to fix that.

all operations seem to still work ok for me.

ok drahn@