Add restart support for metadata upload and download.
At reconnect time we need to walk the list of pending transactions until
we find a metadata opeation related to upload or download. If we find
one we then open the file again (in polled mode) at the correct offset
and let the transactions continue. If the server has lost data after our
write (this should be incredibly rare) then we will fail the open and
user intervention will be required