Fix md download for differential extract.
Add a ct_add_operation_after() api call to add an operation to the list
after the specified operation (instead of at the end of the queue) this
is necessary for recursively downloading the whole md chain for an
extract.

Additionally fix ct_download_next to check in the metadata cache before
downloading a new file. (this is a bit messy and should be cleaned up later)

finally, make sure we set the file to closed again so that we don't accidentally
send a new close when the event loop starts again.