Fix a regression with differential backups.
The previous memory leak fix was wrong in that is freed the ct_basisbackup
pointer a little too early. Since exude zeros the pointer on free, this meant
that the call to ct_setup_write_md setup a metadata header for a non
differential backup despite all the other calculations being differential.

Make the basisbackup parameter one that is only passed through the code as
arguments and make it non-global.

This diff also removes support for the -B option in remote mode. Only
automatic differentials are supported in remote mode (handling ensuring
the basis is in the remote store etc adds extra complexity for very
little gain), -B continues to be supported in local mode