rework the ct operations code somewhat.
instead of having opaquely named arguments (arg1, arg2, arg3, etc) name it so
that each argument to the functions and in the structs have one definition.

As well as making the code somewhat easier to read, it means that they
are typed now.

Adapt all callers to cope with the new naming scheme and with the slightly
different argument order.