Implement a new config directive: md_max_differentials.
If this is set, then after a number of differential backups equal to
this config directory a new level 0 will be carried out automatically.

This is implemented by bumping the metadata version (we support old
versions but older cyphertite instances won't read the new versions) to
store the current level in the header.

Manpage bits contain a much better description than mine from davec@