Add some more fields to the metadata header (this shared the md bump from the last commit)
We put the current working directory and a list of paths backed up into
the header. When performing a differential backup against a md file with
this information we assert that the cwd and the list of paths are the same.
The reason for this is because new directory heirarchies will not be correctly
handled in differential backups (new files and dirs in existing heirarchies will
work fine, of course)

If verbose, print out the full list of errors when dir lists don't match.