Parse new metadata information in md list.
Since we no longer have a static list of strings store the entries we
get back in a rb tree (sorted in lexical order). printing out the new
information comes next. Since we have two lists (before filtering and after),
use a list for the former because we don't need it sorted so no need to balance
the tree.