Support metadata protocol v3.
This is identical to v2 with one change: we byteswap the metadata footer
to network byte order. When I added v2 I missed this, so mixed endian
clients (and clients of a differing byte order than the server) would
not work correctly in metadata {up,down}load.

All big endian clients should be updated following this change.