Add support for crypto secrets upload.
If this option is enabled then we will upload the encrypted secrets file
to the server after login.  If a newer file has been uploaded to the
server then it will be downloaded and used instead, this also applies if
no local secrets file is found.  If this option is disabled then any
secrets files on the server will be deleted the next time cyphertite(1)
is run.

The option is on by default.