Change the way we decide to reconnect slightly.
If we reconnect and the pending queue is empty, then we can assume that
we idled out and the server disconnected us. In that case we set a flag
to attempt reconnect on the first time we need to write. In the case
where a long amount of exists in the local database or a long directory
scan this makes reconnect almost instant.

On the other hand, if we have inflight transactions we reconnect the
same way as before.