Don't fatal out as soon as we fail to create a directory or a file.
Act like tar and print for each failure but continue for the entire backup.
In order to do this properly we need to move the code to create the
files/directories into the file engine instead of the completion engine so we
can know whether or not it succeeded before we start downloading chunks that
we'll have to bin right into /dev/null.

This finds another problem where the global filedescriptor for extract is needed
by more than one thing, so move the fd for each extract file into its fnode.

diff -r is happy with this diff and a backup with no complications.
extracting as a normal user to /usr/bin spews a lot of garbage but gets
to the end.