Reimplement a patch that does the same as a previously backed out diff:
Dont' fatal out as soon as we fail to create a directory or a file.  Act
like tar and print for each failure but continue for the entire backup.

In this case though we just mark a failed open file to be skipped, throw away
any data we recieve for it and don't ask the server for any more.

This will cause more downloading to happen for skipped files (upper
bound is MAXCHUNKSZ * (max queue length - 1)), but does not risk any problems
with racing for hardlink creation.

Another suitable solution would be to do the same as the backed out diff
(open on transaction creation instead of completion, this means having
multiple files open at once up to queuelen / 2), but move link creation
to a second pass between the file extract and the final chmod. This
should also be dependancy free, but after long discussion with drahn we
have agreed not to risk it until this becomes a performance bottleneck
on extract.

ok drahn@