More memleaks found by exude.
md_node is never freed but is only there to have a zeroed fnode for the
checks elsewhere in the statemachine. have it a global and use & instead
of allocating and never freeing it.

Easy memleak of md_download_next() where md_prev can be ignored.

Actualy free the full fnode on extract_close (was mising the string
members) and extract_special (mising frees at all).

Makes exude a lot happier, still a few to go.

ok drahn@