A kinda ugly fix for a problem with partial extracts.
Problem:

ct -xf mybackup "*cyphertite"

this will try and extract files such as the following:
usr/local/bin/cyphertite
usr/local/include/cyphertite/* etc.

However, this means that the archived nodes for usr, usr/local,
usr/local/bin, etc will not be matched when we try to extract them.
leading to ENOENT errors all over the place and most of the backup you
wanted restored missing.

We already deal with this for files, do the same for directories, links
and device nodes. If the initial creation fails with ENOENT, recursively
make all directories leading to that file (using ct_make_full_path),
then try again. if it fails, print a warning and carry on.

Note that this has the following caveats:
1) we don't currently check existing directories or nodes (if we get
EXXIST) to check that they are the right kind of file.

2) The directory permissions won't be the same as in the backup (since
they are not being generated from the backup).

3) Hardlinks may still fail if they point to a non-existant path that
was in the backup but does not match the glob.

ok drahn@