Initial implemention of cyphertitefb(1) the cyphertite file browser.
This is a WIP implementation of a command line application using libedit
that allows a user to browse through a metadata file, seeing where
various versions were backed up. and gives the option to download the
files for viewing. Some bits are a bit rough around the edges but should
be fairly functional.

A fair amount of the code for playing with libedit and doing completion adapted
from sftp(1) or inspired by same (found a couple of bugs while I was doing so,
though).

Some build fixes for linux/freeBSD from davec and thib. Portability
gubbinz for the openbsd glob implementation and packaging changes both
from davec.