Verify the ctfile passed to us.
Restrict the length to 128 bytes (CT_MD_TAG_MAXLEN) in both remote
and local mode. The rational is that we must ensure that the resulting
filename fits into a pathname component (e.g. 255 bytes on UFS/EXTFS)
and allows for future wiggle room.

While here, to help with interoperability do not allow a few
(CT_MD_TAG_REJECTCHRS) special characters to appear in the name also.

Takes care of FS#108.

Discussed with drahn@, OK davec@.