Fix ct allfiles extract
Change CT_MATCH_RB to have the semantics required of allfiles mode. What
we do is that on the first pass through the file to actually extract we
put every file that has -1 shas into the rb tree. on subsequent files we
change the match mode to MATCH_RB, if a file matches it is then removed
from the rb tree, when the tree is empty we are done. If at the end we
have items left in the tree then something bad happened.

Add a configuration knob for allfiles mode, add ct -a and ct -A to turn it on
and off on a per-archive basis.

ok drahn@