Add a facility to have ``local'' transactions.
The problem here is an operation with a large number of small files, especially
extract where we have a transaction for both open and close.

In the cae with a lot of smal files (about 1 chunk per file) you'd allocate a
transaction for open, one for the sha write, and one for close. Only one of
those actually talks to the server, cutting the theoretical throughput by three.

Local transactions allow us to mark operations that don't touch the
server as not taking up one of those precious slots, allowing more work
to be done locally without filling the server queue.

Patch a heavily tweaked version of a proof of concept by drahn@

v2: include a patch from drahn@ to fix a lost tr_trans_id in ct_extract_file.