Make ct_download_md_next provide the cooked (basenamed) filename to download.
right now we were always providing the filename we got from
ct_metadata_check_prev, this was normally a full path and thus we'd try
to open the full path on the server which doesn't work.

Since we were jumping through some hoops to work out when to free these
files simplify this by duplicating the files for the first call (where
we need to keep the filenames for other operations) and just always
free.