Rework the metadata writing paths into a new api: ctfile_write_*
this minimises the amount of the code that needs to know the file format black
magic, instead the api takes filled in fnodes and does the right thing.

Rework ctfile_write and ctfile_parse to not use as much of the ct_xdr.c
global state, and keep their own XDR pointers, direction, and dirnum
counters.