generalise the arguments to a ct_op.
instead of having a fixed list of arguments (localname, remotename, filelist
etc) each operaton takes a void pointer to a structure that contains the
relavent information.

reorganise the remote mode fetching code to not know so much about internals,
taking a callback to the operation that it should trigger at the end.

Some extra cleanup could probably done with the cull global state, but this
patch does not carry that out.