Switch extract and archive over to using openat() and friends
In order to provide more reliable traversal and help prevent symlink
races we open each file relative to the directory it should be in (which
has itself been opened and checked) to prevent dereferencing a symlink
path we didn't expect. For this we need to keep track of which
directories we still have business in and keep open handles to them.

Summary of changes:

Extract now uses the d_name_tree to have a persistent dnode
cache for all backups (the d_num tree is still cleaned up after each file),
permissions are kept in th dnode and fixed up on close (instead of in a second
pass). Similarly upon open diretories are opened, then chmoded if necessary (and
fixed later).

add mkstemp_at() based on OpenBSD's mkstemp but using openat in provided
directory, use this so that we can still do the mkstemp() + rename() dance we
did before.

Fix the archive directory dependancy path to use dnodes exclusively instead of
needing fnodes/flnodes and to do the stat to convert between the two.

Add code to archive to track the postorder directory entries from fts(3)
so we know when to close a directory (when we are done with it). This
however does nto provide close events for paths (e.g. ct -cvf test.ct
a/b/c/d/e, here only 3 would get an event), clean these up at the end of
the archive run.  Extract lacks this information and thus has to compare
paths and walk backwards along the dir list.

Add support for ct -h to follow symlinks, otherwise use O_NOFOLLOW
pretty much everywhere we are not expecting a symlink.