Fix a problem where use of -I did not backup the relavent directories.
due to the format of a ctfile this would mean that if includefile was
"*.c" then ct -I include -cf backup.ct path/to/source_code would have
all c files in the ``root'' directory.

Solve this by running include/exclude processing AFTER we process the
flist into an fnode (instead of at traverse time), this means that we
always have a fully populated directory list in the name tree, so the
on-demand directory writing works as expected.