Rework secrets file upload.
Instead of the current ``automatic'' method (which has innumerable corner cases
that are very tricky to handle) to to semi automatic.

When we generate a secrets file as part of the wizard give the option to
upload it to the server. On every cyphertite operation check that the
secrets file matches what is on the server, it it does not we fatal and
complain.

Implement ctctl secrets {upload,download} commands to mange this further.