Add ctctl secrets generate to generate a new crypto secrets file.
If secrets_upload is set then we also upload the new file.

Now we *do not* generate a key by default if the secrets file is not
found, it must be done explicitly.

Make the config generation wizard generate (and upload) a new secrets file if it
doesn't need to download one.