fix CT_NO_OPENAT with rooted paths on archive.
there was a disconnect between the safename (where / have been stripped if
necessary) and fname (the normal path name), in archive mode we need to access
fname, not sname (annoyingly in extract mode it is the other way around..
*sigh*)

Also, when the openat code was intially written, fnode->fl_fname in
archive mode was changed to be the name without any directory component.
For the openat code this is correct, for !openat we still need the
directory, so put that back in.

Finally, the -C directory (tdir) wasn't being used for the stat in
populate_fnode_from_flist, add that back in.

A further speedup *may* be possible by making flist (which we don't
print) have the tdir in it from the beginning so we don't have to do
this jiggery pokery repeatedly at every other call. That, however, will
come later if at all.

A lot of debugging and searching done by thib@, bug reported by Todd Fries.