Make struct ct_extract_state created by ``extract type'' operations and placed in the global state.
This contains all the state for the ct_file_extract_* functions. Make
ct_attr (-p) part of this state and pass it as part of the extract args.

Add apis to insert and lookup directories in the name tree, and to get
the root directory node.

Pass this state to the NO_OPENAT portability functions too, so that they
can the rootdir from it.