Add cleaner dirname and basename handling.
The dirname and basename functions do not have identical semantics on
all platforms.  On some they modify the original path, and on others
they don't.  This commit adds replacements for dirname (ct_dirname) and
basename (ct_basename) that ensure the original path is never modified
and are also thread safe.  They return a newly exude allocated string
that must be freed with e_free.

Suggestions, nits, and fixes from oga@, drahn@, and thib@.