Add an initial libct (libcyphertite) directory and move most of the code over to it.
The api in this library is far from final, but this gives us somewhere to work
from. Documentation is in the works.

To keep things non interactive move the logic from
{upload,download}_secrets file back into ctctl.  This means that we have
no api calls (other than those supposed to be so) that are user
interactive in the library. get_anser, get_password etc are kept as
helpers.