Add a framework to allow the event loop to commit suicide.
ct_fatal() when called sets up an event that will kill everything, free
all we can then return from the event loop. To be used to return an
error code from the event loop instead of killing the library caller.

There are some cases where we can't free data (ctfile_extract we will
probably leak the fnode, and any arguments that belong to the
ct_operation framework as of yet are not cleaned up) But generally most things
should be cleaned up:

If fatal we do not called the tr_complete callback on transactions, but we do
call tr_cleanup. We also wakeup all threads of control so they can clean up
their own local state.