Make ctfile_get_previous be able to return an error.
Additionally add paremeters to pass in the cache directory in question. The
return value is still *without* the cache directory since callers tend to want
it that way.

Fixes a bug where cull was not looking in the right place for files and thus teh
dependancy tracking was not kicking in. If we fail to open the file this is
fatal for cull. Better safe than accidentally deleting all your data.