Make ct -mef validate files can be deleted.
To avoid removing files with dependancies  ct -mef now checks the files
to be deleted are not depended on by any other file stored on the
server.

The -f argument is now a glob to enable deleting whole parts of a ctfile
chain at once. If a file in invalid for now we just skip it for
dependancy purposes.

This could be optimised a bit by only searching the list of tags that the
glob matches instead of all ctfiles, but for now better safe than sorry.