Populate the total bytes statistic for extract
So that we can provide proper progress statistics, we need to know the
total amount of data to be extracted. The ctfile format has no
information for that, so we first of all do a dry-run extract wherewe
total up the size of all the files to be extracted (for this purpose
anything local, e.g. device nodes, symlinks, directories and hardlinks
do not count). AFter this isdone we continue with the main extract.

Note that this can mean that quickly changing non-allfiles backups can
have very large totals due to the sequential extract model used.
allfiles backups only extract the minimal so won't do this.