ctctl secrets generate is dangerous operation.
replacing the current secrets file will invalidate all data already stored on
the server. In order to catch most invalid uses of this we have the following
lines of defence:

1) (already there) if a crypto secrets file already exists, fatal out asking the
user to remove it first.
2) if any files are stored on the server, fatal out with an explanatory message
asking the user to remove all of these and re-run (they are invalid anyway,
forcing the user to do the removal is the confirmation we require that you mean
to do this)
3) if this does happen. trim the ctfile_cachedir and unlink the local database
because any information held therein is about to be completely invalidated.