make ct -me take pattern as a filelist.
this means taht the usage is now:

ct -me  glob1 glob2 glob3

instead of

ct -mef <file>

this runs around a bug where we couldn't glob because -f arguments are
checked for being valid ctfile names.

There is debate about whether we should explicitly ban "*" as a pattern,
but for now this is not done.

Update manpage for same.