Handle unsolicited cull messages
If we get an unsolicited newdbgenid message it means that a cull
occurred while we were sending shas. In this event we set the localdb to
the new genid, run exists on the whole backup (aftercompletion) to
ensure all shas are marked as precious, if one does not exist we re-run
the archive.

Discussed heavily with drahn