Implement keep-alive timeout.
This commit implements a keep-alive timeout so libevent always has an
event waiting while file processing is taking place to ensure it doesn't
exit prematurely after a disconnect.

It is possible that no communications will take place with the server
for quite some time during file processing if the data is already on the
server and the local database already has this information cached.  In
such a scenario, ct will disconnect from the server until it has data to
send.  Without this change, this could result in the event loop
finishing early when there are no more events.

Issue reported by miken@.  Help from oga@ and drahn@ in tracking down
the issue.