Rework how we hadle paths somewhat.
the addition of 3factor incremental selections has made it very important that
paths we get out of the ctfile are the same as processed, otherwise we will fail
the presence check.

- firstly, -P no longer works for archive, all archives keep full paths.
extracts default to stripping slashes still though.

- secondly, don't use safenames anywhere, they are non isomorphic
(stripping leading ../ compeltely without replacement). change to
storing the shortname in fl_name and the full name in fl_sname (the same
as extract).  fl_fname remains only for temp names in extract (to be
renamed in a subsequent commit).

now incrementals works properly again for absolute paths.