remove eat_double_dots()
it is doing magic that we don't really want (we record the real paths given to
use and strip later if we are stripping slashes). This removes the spurious .s
that davec was finding.

Add an extra check for a path separator in backup_prefix (another
function i would like to consign to the depths) so that we don't get
more spurious . since dirname("a") is ".".

discussed with drahn