Fix hardlink failures with allfiles.
Replace the match_rb ct_match structure for pending allfiles extracts (which are
in a previous level) with a custom rb tree that keeps two things:

1) permissions at time of backup (this could vary from the time the fiel was
stored), we use this on extarct instead of those in the ctfile.
2) a list of hardlinks to that file.

When we are to make a hardlink, check if we have a file for the destination in
the pending list, if we do then don't even try and make it now, put it on the
pending list. When the file is closed we also make any pending hardlinks.

The names for the apis in this commit are pretty sucky, but the code works. the
apis should be fixed namewise soonish.

additonally we should switch the extract size calculation over to using this
code and remove match_rb.

ok drahn@