Add escape documentation to man page and conf file.
The config file parsing logic has changed to support use of characters
that would ordinarily have special meaning if they are escaped with a
backslash.  This commit updates the man page to reflect this change and
adds a comment to generated cyphertite.conf files so a user that chooses
to manual edit it in the future is aware of the requirement.