Move to using an internal version of fts.
This is a copy of the openbsd fts() code with all flags and function
names appended with ct_. Soon to be slightly modified, for now the only
changes are the name changes and three small variable settings to handle
-Wuninitialized-somtimes false positives with clang.  oga@

Enable large file support for 32bit Linux.

Add initial test case.

ok davec@