fun with disconnects.
Handle errors in ct_socket a little more comprehensively. Treat empty read/write
and all errors that are no EINTR or EAGAIN as meaning the connection is dead.
When we get them give the universal signal for connection died by calling the
appropriate callback with NULLs.

Should make disconnect a mite more robust.