Make fnodes be refcounted.
Use this fact to fix a bug where we fatal at an inopportune time. There
is no way to know whether or not a transaction is acting on the fnode
(which may be being acted on by the sha or encrypt threads) so we can't
free it from the file thread explicitly, ever. We still don't want to
leak memory though, so we refcount  so that every transaction that
references that fnode has a reference, as well as the file thread while
processing it. This way we will free the fnode when all are gone without leaks
*or* crashes.

Note that the changes removed the fnode variable from ct_extract, this
was because it was quite hard to follow the data dependencies through
the loop without this change.

Bug found and extra testing by wenzo.