fix bug in previous.
If we call ct_get_next_fnode twice at the end then we will restart the looping
thogh all files and hit a number of possible asserts the second time around
(already allocated dnums is a popular one).

Add some special handling so that when we detect we are done, we are
done and we don't try and do any mroe.

This function needs a cleanup, it is getting pretty nasty.