Fix ctfb when / is explicitly backed up.
The problem here is that the code handling the version tree all assumes
that no entry has the name of the path separator. If / itself is backed
up this is untrue. The solution we use is when generating the version
tree, we explicitly detect this case and move all children of the /
directory to be parents of the `root' node in the tree, thus restoring
the previous assumption.

Closes #41

Problem reported by Björn Ketelaars; thank you!