Unconditionally replace certs in download_and_decode_certs()
There were issues where if someone had an old config and tried to set up
a new account, we would attempt to download certs and already have an
existing ca file. The checks that run stat on certs file in this
function would then *fail* cert download. Remove all of the stat code
from this function and just assume that if someone calls it they want
the certs downloaded from the website in to the paths in questions. No
ifs, no buts.

Discussed with wenzo@