Add support for cyphertite.com/update.php
Adds code similar to ct_certs.c that uses curl to pull down an xml with
current cyphertite version information.  The xml is to be parsed by the
caller.

A few private callback functions we set for curl in ct_certs.c have been made public so they can be used by other portions of the code.

From wenzo.