Adds cve_exists to version tree nodes
Some clients of libct (such as the Windows GUI) allow the user to browse
individual ctfiles. If the user backs up a directory, deletes one file,
and then backs the directory up again, the expected behavior is that this
file will *only* be visible when browsing the level 0 ctfile, and *not* the
level 1.

The challenge is that the level 1 version tree still (indirectly) contains
a reference to the deleted file as a prior version in the incremental
chain. This means that the full version tree *must* contain the file as a
node. However, in previous implementations, the nodes did not indicate the
ctfile the node was first found in. As such, the browser had to assume
that every node existed at the time of the backup, causing it to falsely
show files that were not actually in the requested ctfile.

Since the version tree is built off of an input ctfile path, this commit
simply adds a flag to each node that indicates whether the node was found
in the input ctfile (cve_exists = 1) or only found in a prior ctfile
(cve_exists = 0).